IdeaInspiration/Sources/Community/PromptBoxSource/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prismq-idea-sources-community-promptboxsource"
version = "1.0.0"
description = "PrismQ module for collecting user-submitted prompts and ideas"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [{name = "PrismQ", email = "dev@prismq.com"}]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "sqlalchemy>=2.0.0",
    "vaderSentiment>=3.3.2",
]

[project.scripts]
prompt-box-source = "src.cli:main"

[tool.setuptools]
packages = ["src"]
